asklvkaszus-backend/docker/requirements.txt
bleach==6.0.0
Blinder==0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Deprecated==1.2.14
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Limiter==3.3.1
gunicorn==21.2.0
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.1.0
ordered-set==4.1.0
packaging==23.1
protobuf==4.21.12
Pygments==2.15.1
requests==2.31.0
rich==13.4.2
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
redis==4.6.0